+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.9.1
 
 * Expose `Messages.PlatformMarker.InfoWindow.toJson` for pigeon conversion.
+* Updates minimum supported SDK version to Flutter 3.19/Dart 3.3.
 
 ## 2.9.0