+++ packages/pointer_interceptor/pointer_interceptor/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+* Updates minimum supported SDK version to Flutter 3.19/Dart 3.3.
+
 ## 0.10.1+2
 
 * Adds performance warning about using multiple pointer interceptors on iOS.